refactor(rcml+shopify): extract createEmailTemplate helper + collapse abandoned-cart to a pure template
New @rule-io/rcml helper `createEmailTemplate<TCopy, TContext>({baseUrl,
templatePath, copyPath})` bundles the full email-template render
pipeline: load XML + JSON copy at construction, merge caller copy
override, project `theme.images.logo?.url` / `theme.links` into the
compile context, compile, xmlToRcml, applyTheme. Its home is rcml
(not templates) because it needs EmailTheme + applyTheme + xmlToRcml.

Abandoned-cart becomes the canonical "four-file template" shape:

  abandoned-cart/
  ├── abandoned-cart.xml          # layout
  ├── abandoned-cart-copy.json    # default copy
  ├── abandoned-cart.ts           # context/copy types + factory
  └── abandoned-cart.test.ts

The factory body is a single createEmailTemplate(...) call.
AbandonedCartTemplate / AbandonedCartRenderOptions are kept as type
aliases over the generic EmailTemplate / EmailTemplateRenderArgs —
zero public-API break for callers of the factory.

Drop the config-driven createAbandonedCartEmail builder + its test +
the shopify-abandoned-cart automation entry. Downstream code will
build the abandoned-cart context explicitly (via customField /
loopValue + client-fetched entities) and call
createAbandonedCartTemplate().render({context, theme}) directly.

- new: packages/rcml/src/email/create-email-template.{ts,test.ts} +
  two fixture files (xml + json)
- rcml index.ts: re-export createEmailTemplate et al.
- vendor-shopify: rename create-abandoned-cart-template.ts →
  abandoned-cart.ts (git mv, history preserved); collapse factory
  body to a single createEmailTemplate call
- vendor-shopify: delete config-driven abandoned-cart.ts + test
- vendor-shopify index.ts: drop createAbandonedCartEmail /
  AbandonedCartConfig exports
- vendor-shopify automations.ts: drop import + ABANDONED_CART_TEXT +
  the shopify-abandoned-cart automation entry (now 4 automations)
- tests/shopify.test.ts: drop abandoned-cart-specific assertion;
  adjust count from 5 to 4
- JSDoc cleanups in core (automation-configs-v2.ts) + client
  (client.ts) + sdk README

rcml: 1077 tests (+7 new). templates: 208 tests (unchanged).
vendor-shopify: 75 tests (was 84; -8 config-driven tests, -1
automation-preset test). All green; types clean across the affected
packages.
