+++ packages/rcml/src/email/create-email-template-fixture.json
@@ -1,0 +1,3 @@
+{
+  "greeting": "Hello {{name}}"
+}
