+++ packages/rcml/src/email/index.ts
@@ -22,3 +22,4 @@ export * from './create-theme.js'
 export * from './apply-theme.js'
 export * from './get-theme.js'
 export * from './brand-elements.js'
+export * from './create-email-template.js'
